+++ lib/output.php
@@ -25,8 +25,11 @@ add_action(
 	'wp_enqueue_scripts',
 	function () {
 
-		$color_link   = get_theme_mod( 'genesis_advanced_link_color', genesis_advanced_customizer_get_default_link_color() );
-		$color_accent = get_theme_mod( 'genesis_advanced_accent_color', genesis_advanced_customizer_get_default_accent_color() );
+		/* Locate the config file */
+		$appearance = genesis_get_config( 'appearance' );
+
+		$color_link   = get_theme_mod( 'genesis_advanced_link_color', $appearance['default-colors']['link'] );
+		$color_accent = get_theme_mod( 'genesis_advanced_accent_color', $appearance['default-colors']['accent'] );
 		$logo         = wp_get_attachment_image_src( get_theme_mod( 'custom_logo' ), 'full' );
 
 		if ( $logo ) {
@@ -36,7 +39,7 @@ add_action(
 		$css = '';
 
 		/* Output link color inline css */
-		$css .= ( genesis_advanced_customizer_get_default_link_color() !== $color_link ) ? sprintf(
+		$css .= ( $appearance['default-colors']['link'] !== $color_link ) ? sprintf(
 			'
 			a,
 			a:focus,
@@ -48,7 +51,7 @@ add_action(
 		) : '';
 
 		/* Output accent color inline css */
-		$css .= ( genesis_advanced_customizer_get_default_accent_color() !== $color_accent ) ? sprintf(
+		$css .= ( $appearance['default-colors']['accent'] !== $color_accent ) ? sprintf(
 			'
 			.button.accent,
 			a.button.accent,