+++ lib/setup.php
@@ -46,3 +46,24 @@ add_action(
 	},
 	9
 );
+
+add_action(
+	'after_setup_theme',
+	'genesis_advanced_content_width',
+	0
+);
+
+/**
+ * Set the maximum allowed width for any embedded content
+ *
+ * @since 3.4.0
+ */
+function genesis_advanced_content_width() {
+
+	/* Locate the config file */
+	$appearance = genesis_get_config( 'appearance' );
+
+	// phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedVariableFound -- See https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/issues/924
+	$GLOBALS['content_width'] = apply_filters( 'genesis_advanced_content_width', $appearance['content-width'] );
+
+}